chore: extract the ship into multiple files
